==> Notebook/01_CleaningDataset.ipynb <==
##**Mounting Google Drive**

from google.colab import drive
drive.mount('/content/drive')

##**Import Packages**

import numpy as np
import pandas as pd

###*LIAR*

## Path of Dataset
trainFile = '/content/drive/MyDrive/BTP/Dataset/LIAR/train.tsv'
testFile = '/content/drive/MyDrive/BTP/Dataset/LIAR/test.tsv'
valFile = '/content/drive/MyDrive/BTP/Dataset/LIAR/valid.tsv'

## Name of headers
headers = ["json ID", "label", "statement", "subject", "speaker", "job title", "state", "party", 
           "barely true", "false", "half true", "mostly true", "pants on fire", "los"]

## Loading Data
traindata = pd.read_csv(trainFile, delimiter='\t', encoding='utf-8', names=headers)
testdata = pd.read_csv(testFile, delimiter='\t', encoding='utf-8', names=headers)
valdata = pd.read_csv(valFile, delimiter='\t', encoding='utf-8', names=headers)

## Concatenation
dataset1 = pd.concat([traindata, testdata, valdata])

## Labels encoding
label_dic = {
            "true" : 1,
            "mostly-true" : 1,
            "half-true" : 1,
            "barely-true" : 0,
            "false" : 0,
            "pants-fire" : 0
        }

## Binary conversion
dataset1['label'] = dataset1['label'].map(label_dic)

## Removing Unnecessary Columns
removed_col = ['json ID', 'barely true', 'false', 'half true', 'mostly true', 'pants on fire', 'los', 
               'party', 'state', 'subject', 'job title', 'speaker']
dataset1.drop(removed_col, axis='columns', inplace=True)

## Saving 
dataset1.to_csv("/content/drive/MyDrive/BTP/Files/LIAR_Binary.csv", encoding='utf-8', index=False)

dataset1.head()

dataset1.describe()

###*ISOT*

fakeFile = '/content/drive/MyDrive/BTP/Dataset/ISOT/Fake.csv'
trueFile = '/content/drive/MyDrive/BTP/Dataset/ISOT/True.csv'

fakedata = pd.read_csv(fakeFile, encoding='utf-8')
truedata = pd.read_csv(trueFile, encoding='utf-8')

## Adding Labels
fakedata['label'] = "false"
truedata['label'] = "true"

dataset2 = pd.concat([fakedata, truedata])

## Renaming Column
dataset2.rename(columns = {'text':'statement'}, inplace = True)

label_dic = {
            "true" : 1,
            "false" : 0,
        }

dataset2['label'] = dataset2['label'].map(label_dic)

removed_col = ['subject', 'date', 'title']
dataset2.drop(removed_col, axis='columns', inplace=True)
dataset2 = dataset2[['label','statement']]

dataset2.to_csv("/content/drive/MyDrive/BTP/Files/ISOT_Binary.csv", encoding='utf-8', index=False)

dataset2.head()

dataset2.describe()

###*POLITIFACT*

File = '/content/drive/MyDrive/BTP/Dataset/POLITIFACT/politifact.csv'

dataset3 = pd.read_csv(File, encoding='utf-8')

## Dropping Unncessary Values
dataset3.drop(dataset3.index[dataset3['fact'] == 'half-flip'], inplace=True)
dataset3.drop(dataset3.index[dataset3['fact'] == 'no-flip'], inplace=True)
dataset3.drop(dataset3.index[dataset3['fact'] == 'full-flop'], inplace=True)

dataset3.rename(columns = {'fact':'label', 'curators_article_title' : 'title', 'curator_complete_article' : 'statement'}, inplace = True)

label_dic = {
            "true" : 1,
            "mostly-true" : 1,
            "half-true" : 1,
            "barely-true" : 0,
            "false" : 0,
            "pants-fire" : 0
        }

dataset3['label'] = dataset3['label'].map(label_dic)

removed_col = ['sources', 'sources_dates', 'sources_post_location', 'sources_quote', 'curator_name', 'curated_date', 
               'sources_url', 'curator_tags', 'Unnamed: 0', 'title']
dataset3.drop(removed_col, axis='columns', inplace=True)

## Removing Nan Values
dataset3 = dataset3[dataset3['statement'].notna()]

dataset3.to_csv("/content/drive/MyDrive/BTP/Files/POLITIFACT_Binary.csv", encoding='utf-8', index=False)

dataset3.head()

dataset3.describe()

dataset = pd.concat([dataset1, dataset2, dataset3], ignore_index = True)
dataset.to_csv("/content/drive/MyDrive/BTP/Files/dataset.csv", encoding='utf-8', index=False)
dataset.head()

dataset.describe()

dataset.isna().sum()
