==> Notebook/04_Model3.ipynb <==
##**Mounting Google Drive**

from google.colab import drive
drive.mount('/content/drive')

##**Import Packages**

import numpy as np
import pandas as pd
from sklearn.ensemble import  RandomForestClassifier, ExtraTreesClassifier, BaggingClassifier, GradientBoostingClassifier
from xgboost import XGBClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC

##**Loading Dataset**

train_data = pd.read_csv("/content/drive/MyDrive/BTP/Files/train_data.csv", encoding='utf-8')
test_data = pd.read_csv("/content/drive/MyDrive/BTP/Files/test_data.csv", encoding='utf-8')

Y_train = train_data['label'].values
Y_test = test_data['label'].values

X_train = train_data.drop(['label','statement'], axis = 1).values
X_test = test_data.drop(['label','statement'], axis = 1).values

X_train.shape

##**Models**

Logistic Regession

lr = LogisticRegression()
lr.fit(X_train, Y_train)

print("Training Accuracy: ", lr.score(X_train, Y_train))
print("Testing AccuracY: ", lr.score(X_test, Y_test))

Naive Bayes

nb = GaussianNB()
nb.fit(X_train, Y_train)

print("Training Accuracy: ", nb.score(X_train, Y_train))
print("Testing AccuracY: ", nb.score(X_test, Y_test))

SVM

svm = SVC()
svm.fit(X_train, Y_train)

print("Training Accuracy: ", svm.score(X_train, Y_train))
print("Testing AccuracY: ", svm.score(X_test, Y_test))

Random Forest

rf = RandomForestClassifier()
rf.fit(X_train, Y_train)

print("Training Accuracy: ", rf.score(X_train, Y_train))
print("Testing AccuracY: ", rf.score(X_test, Y_test))

Gradient Boosting

gb = GradientBoostingClassifier()
gb.fit(X_train, Y_train)

print("Training Accuracy: ", gb.score(X_train, Y_train))
print("Testing AccuracY: ", gb.score(X_test, Y_test))

XGBoost Classifier

xgb = XGBClassifier()
xgb.fit(X_train, Y_train)

print("Training Accuracy: ", xgb.score(X_train, Y_train))
print("Testing AccuracY: ", xgb.score(X_test, Y_test))

Bagging With Decision Tree

bag_dt = BaggingClassifier(
    base_estimator = DecisionTreeClassifier()
)
bag_dt.fit(X_train, Y_train)

print("Training Accuracy: ", bag_dt.score(X_train, Y_train))
print("Testing AccuracY: ", bag_dt.score(X_test, Y_test))
