==> Notebook/05_Ensemble.ipynb <==
##**Mounting Google Drive**

from google.colab import drive
drive.mount('/content/drive')

##**Import Packages**

import numpy as np
import pandas as pd

##**Loading Predictions**

model1_train_pred = pd.read_csv("/content/drive/MyDrive/BTP/Results/model1_train_pred.csv", encoding='utf-8')
model2_train_pred = pd.read_csv("/content/drive/MyDrive/BTP/Results/model2_train_pred.csv", encoding='utf-8')
model1_test_pred = pd.read_csv("/content/drive/MyDrive/BTP/Results/model1_test_pred.csv", encoding='utf-8')
model2_test_pred = pd.read_csv("/content/drive/MyDrive/BTP/Results/model2_test_pred.csv", encoding='utf-8')
test_data = pd.read_csv("/content/drive/MyDrive/BTP/Files/test_data.csv", encoding='utf-8')
train_data = pd.read_csv("/content/drive/MyDrive/BTP/Files/train_data.csv", encoding='utf-8')

##**Calculating Accuracy**

def Accuracy(pred1, pred2, pred):
  cnt = 0
  for i in range(len(pred1)):
    val = ((1.7*pred1[i] + pred2[i])/2.7)
    if(val > 0.5):
      val = 1
    else:
      val = 0

    if val == pred[i]:
      cnt += 1

  return (cnt / len(pred1))

Accuracy(model1_train_pred['predictions'].values, model2_train_pred['predictions'].values, train_data['label'].values)

Accuracy(model1_test_pred['predictions'].values, model2_test_pred['predictions'].values, test_data['label'].values)
